Allowing to let `complete' override `complete'
The general rule in caching was to only let newer statements override older
statements and only if the existing statement (concerning a certain activity
and user combination) did not have the completed verb. This is order to prevent
new launched statements to override completed ones. However this is now
extended to make sure that if a new completed statement comes in for that
context (i.e. the activity has been completed another time) then the latest
result is kept.